+++ NEWS.md
@@ -1,4 +1,5 @@
 # tablerDash 0.2.0.9000
+- add preloader in `tablerDashPage()`
 - add `tablerList()` and `tablerListItem()`
 - add `tablerTable()` and `tablerTableItem()`